ansible-streamlit-app/requirements.txt
streamlit==1.28.1
ansible==8.5.0
ansible-core==2.15.5
pyyaml==6.0.1
jinja2==3.1.2
paramiko==3.3.1
watchdog==3.0.0 